Merge: Merge branch '3.5.x' into 3.6.x

+++ src/backend/job-manage/api-job-manage/src/main/java/com/tencent/bk/job/manage/model/web/vo/globalsetting/FileUploadSettingVO.java
@@ -41,7 +41,7 @@ import java.util.List;
 @Data
 public class FileUploadSettingVO {
     @ApiModelProperty("数量")
-    private Long amount;
+    private Float amount;
     @ApiModelProperty("单位")
     private String unit;
     @ApiModelProperty("限制模式，0:禁止范围，1：允许范围")
